% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_start_and_stop_codons.R
\name{add_start_and_stop_codons}
\alias{add_start_and_stop_codons}
\title{Add start and stop codons}
\usage{
add_start_and_stop_codons(nt_sequence)
}
\arguments{
\item{nt_sequence}{Vector of nucleotides.}
}
\value{
Vector of nucleotides with start and stop c("T","A","G") codons
prepended and appended, respectively. An error will be returned if the
number of nucleotides is not a multiple of 3.
}
\description{
Add start and stop codons
}
\examples{
add_start_and_stop_codons(c("A","T","A"))
}
