% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deletion_mutator}
\alias{deletion_mutator}
\title{Randomly delete nucleotides in a sequence}
\usage{
deletion_mutator(nt_sequence, deletion_rate_A = 0.05,
  deletion_rate_T = 0.05, deletion_rate_C = 0.05, deletion_rate_G = 0.05)
}
\arguments{
\item{nt_sequence}{A vector of nucleotides, i.e. c("C","A","T","G","A").}

\item{deletion_rate_A}{The probability of an "A" deletion occurring.}

\item{deletion_rate_T}{The probability of an "T" deletion occurring.}

\item{deletion_rate_C}{The probability of an "C" deletion occurring.}

\item{deletion_rate_G}{The probability of an "G" deletion occurring.}
}
\value{
A vector of nucleotides, i.e. c("C","T","G","A"), with deleted
nucleotides omitted.
}
\description{
Randomly delete nucleotides in a sequence
}
\details{
Pass in a vector of nucleotides, the probability of single
nucleotide deletions for each of the four nucleotides, and return a new
vector of nucleotides.
}
\examples{
set.seed(1)
deletion_mutator(
c("C","A","T","G","A"),
deletion_rate_A = 0.05,
deletion_rate_T = 0.05,
deletion_rate_C = 0.05,
deletion_rate_G = 0.05
)

}
