% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{.compact}
\alias{.compact}
\title{Compact a list removing NULL elements}
\usage{
.compact(x)
}
\arguments{
\item{x}{A list possibly containing \code{NULL} elements.}
}
\value{
The same list with all \code{NULL} elements removed.
}
\description{
Compact a list removing NULL elements
}
\keyword{internal}
