% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{.evo_path}
\alias{.evo_path}
\title{Build internal API path}
\usage{
.evo_path(...)
}
\arguments{
\item{...}{Character path segments.}
}
\value{
A single character scalar with segments concatenated by "/".
}
\description{
Build internal API path
}
\keyword{internal}
