% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{.normalize_media}
\alias{.normalize_media}
\title{Normalise media input (URL, file path, base64, data-URI)}
\usage{
.normalize_media(x)
}
\arguments{
\item{x}{The media input to normalise.}
}
\value{
Character string suitable for the API body (URL or base64).
}
\description{
Normalise media input (URL, file path, base64, data-URI)
}
\keyword{internal}
