% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_start_codon.R
\name{find_start_codon}
\alias{find_start_codon}
\title{Find the first start codon in a sequence.}
\usage{
find_start_codon(nt_sequence)
}
\arguments{
\item{nt_sequence}{A vector of nucleotides, i.e. c("C","A","T","G","A").}
}
\value{
Start codon site or NA if not found.
}
\description{
This function will return the index in a vector of nucleotides
of the first start codon.
}
\examples{
find_start_codon(
c("A", "A", "T", "G", "C", "A", "A", "T", "C", "A", "T", "A", "G", "A")
)

}
