% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequence.R
\name{generate_sequence}
\alias{generate_sequence}
\title{Generate a DNA sequence.}
\usage{
generate_sequence(n = 100, start_codon = TRUE, stop_codon = TRUE,
  probabilities = c(0.25, 0.25, 0.25, 0.25),
  allow_internal_stop_codons = FALSE)
}
\arguments{
\item{n}{Number of codons in the sequence.}

\item{start_codon}{TRUE/FALSE, whether a start codon should be included.}

\item{stop_codon}{TRUE/FALSE, whether a stop codon should be included.}

\item{probabilities}{Probabilities for A, T, C, and G, respectively.}

\item{allow_internal_stop_codons}{TRUE/FALSE, whether stop codons should be
generated in the sequence (not including the terminus if stop_codon is TRUE).}
}
\value{
A vector of nucleotides
}
\description{
Generate a DNA sequence.
}
\examples{
generate_sequence(100, start_codon = TRUE, stop_codon = TRUE)
}
