% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertion_mutator.R
\name{insertion_mutator}
\alias{insertion_mutator}
\title{Iterate through the nucleotides and randomly insert nucleotides.}
\usage{
insertion_mutator(nt_sequence, insertion_rate_A = 0.05,
  insertion_rate_T = 0.05, insertion_rate_C = 0.05,
  insertion_rate_G = 0.05)
}
\arguments{
\item{nt_sequence}{A vector of nucleotides, i.e. c("C","A","T","G","A").}

\item{insertion_rate_A}{The probability of an "A" insertion occurring.}

\item{insertion_rate_T}{The probability of an "T" insertion occurring.}

\item{insertion_rate_C}{The probability of an "C" insertion occurring.}

\item{insertion_rate_G}{The probability of an "G" insertion occurring.}
}
\value{
A vector of nucleotides, i.e. c("C","A","T","G","A"), with any
insertions made included.
}
\description{
Pass in a vector of nucleotides, the probability of any given
nucleotide being inserted, and get back a vector with insertion mutations.
}
\examples{
set.seed(1)
insertion_mutator(
c("C","A","T","G","A"),
insertion_rate_A = 0.05,
insertion_rate_T = 0.05,
insertion_rate_C = 0.05,
insertion_rate_G = 0.05
)

}
