% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{substitution_mutator}
\alias{substitution_mutator}
\title{Perform substitution mutations on a nucelotide sequence.}
\usage{
substitution_mutator(nt_sequence, sub_rate_A_to_T = 0.05,
  sub_rate_A_to_C = 0.05, sub_rate_A_to_G = 0.05, sub_rate_T_to_A = 0.05,
  sub_rate_T_to_C = 0.05, sub_rate_T_to_G = 0.05, sub_rate_C_to_A = 0.05,
  sub_rate_C_to_T = 0.05, sub_rate_C_to_G = 0.05, sub_rate_G_to_A = 0.05,
  sub_rate_G_to_T = 0.05, sub_rate_G_to_C = 0.05)
}
\arguments{
\item{nt_sequence}{A vector of nucleotides, i.e. c("C","A","T","G","A").}

\item{sub_rate_A_to_T}{The probability of an "A" becoming a "T".}

\item{sub_rate_A_to_C}{The probability of an "A" becoming a "C".}

\item{sub_rate_A_to_G}{The probability of an "A" becoming a "G".}

\item{sub_rate_T_to_A}{The probability of an "T" becoming a "A".}

\item{sub_rate_T_to_C}{The probability of an "T" becoming a "C".}

\item{sub_rate_T_to_G}{The probability of an "T" becoming a "G".}

\item{sub_rate_C_to_A}{The probability of an "C" becoming a "A".}

\item{sub_rate_C_to_T}{The probability of an "C" becoming a "T".}

\item{sub_rate_C_to_G}{The probability of an "C" becoming a "G".}

\item{sub_rate_G_to_A}{The probability of an "G" becoming a "A".}

\item{sub_rate_G_to_T}{The probability of an "G" becoming a "T".}

\item{sub_rate_G_to_C}{The probability of an "G" becoming a "C".}
}
\value{
A vector of nucleotides, i.e. c("C","A","A","G","A"), with
subsitution mutations included.
}
\description{
Perform substitution mutations on a nucelotide sequence.
}
\details{
Pass in a vector of nucleotides and the probability of each type of
 substitution mutation. Get back a mutated vector.
}
\examples{
set.seed(1)
substitution_mutator(
c("C","A","T","G","A")
)

}
