% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate a DNA sequence into an amino acid sequence.}
\usage{
translate(nt_sequence, check_start_and_stop = TRUE)
}
\arguments{
\item{nt_sequence}{An in-frame vector of nucleotides, i.e. c("A","T","G","T"
,"A","G"). The first and last codons should be start and stop codons.}

\item{check_start_and_stop}{Logical indicating whether the first and last
codons need to be start and stop codons, respectively.}
}
\value{
A vector of amino acids.
}
\description{
Translate a vector of DNA nucleotides sequence into an amino
acid sequence.
}
\examples{
translate(
c("A","T","G","T","A","G")
)

}
