% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stop_codon.R
\name{find_stop_codon}
\alias{find_stop_codon}
\title{Find the first in-frame stop codon}
\usage{
find_stop_codon(nt_sequence)
}
\arguments{
\item{nt_sequence}{A vector of nucleotides, i.e. c("A","T","G","A", ...) in
which the first codon is the start codon.}
}
\value{
Index of the first in-frame stop codon (or NA if not found).
}
\description{
This function will return the index in a vector of nucleotides
of the first stop codon. nt_sequence is assumed to be in-frame.
}
\examples{
find_stop_codon(
c("A", "A", "T", "G", "C", "A", "A", "T", "C", "A", "T", "A", "G", "A")
)

}
