% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{send_reaction}
\alias{send_reaction}
\title{React to a message}
\usage{
send_reaction(client, key, reaction, verbose = FALSE)
}
\arguments{
\item{client}{An \code{\link[=evo_client]{evo_client()}} object.}

\item{key}{List with \code{remoteJid}, \code{fromMe}, \code{id} of the target message.}

\item{reaction}{Emoji like \code{"\\xF0\\x9F\\x98\\x81"}.}

\item{verbose}{Logical. If TRUE, logs request/response details with \code{cli} and enables \code{req_verbose()}.}
}
\description{
React to a message
}
