% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.R
\name{send_sticker}
\alias{send_sticker}
\title{Send a sticker}
\usage{
send_sticker(client, number, sticker, delay = NULL, verbose = FALSE)
}
\arguments{
\item{client}{An \code{\link[=evo_client]{evo_client()}} object.}

\item{number}{Character. Recipient in E.164 format (e.g., \code{"+5581999..."}).}

\item{sticker}{URL or base64 image.}

\item{delay}{Integer (ms). Optional presence delay before sending.}

\item{verbose}{Logical. If TRUE, logs request/response details with \code{cli} and enables \code{req_verbose()}.}
}
\description{
Send a sticker
}
